fixing structure of project: renaming templates
