+++ mocks/offers.js
@@ -1,5 +1,5 @@
 import {EventType} from '../src/const.js';
-import {generateRandomInt, pickRandomElement} from '../utils/random.js';
+import {generateRandomInt, pickRandomElement} from '../src/utils/random.js';
 
 const offerDetails = [`Rent a car`, `Good option`, `Perfect option`, `Ideal option`, `Improve class`, `Take a boost`];
