+++ src/main.js
@@ -1,15 +1,15 @@
-import {createTripSummaryTemplate} from "./view/trip-summary.js";
-import {createTripPathTemplate} from "./view/trip-path.js";
-import {createTripCostTemplate} from "./view/trip-cost.js";
-import {createTripTabsTemplate} from "./view/trip-tabs.js";
-import {createEventFilterTemplate} from "./view/event-filter.js";
-import {createEventSorterTemplate} from "./view/event-sorter.js";
-import {createEventFormTemplate} from "./view/event-form.js";
-import {createEventDayTemplate} from "./view/event-day.js";
-import {createEventItemTemplate} from "./view/event-item.js";
-import {generateEvent} from "../mocks/event.js";
-import {generateDestinations} from "../mocks/destinations.js";
-import {generateOffers} from "../mocks/offers.js";
+import {createTripSummaryTemplate} from './view/trip-summary.js';
+import {createTripPathTemplate} from './view/trip-path.js';
+import {createTripCostTemplate} from './view/trip-cost.js';
+import {createTripTabsTemplate} from './view/trip-tabs.js';
+import {createEventFilterTemplate} from './view/event-filter.js';
+import {createEventSorterTemplate} from './view/event-sorter.js';
+import {createEventEditorTemplate} from './view/event-editor.js';
+import {createEventDayTemplate} from './view/event-day.js';
+import {createEventPointTemplate} from './view/event-point.js';
+import {generateEvent} from '../mocks/event.js';
+import {generateDestinations} from '../mocks/destinations.js';
+import {generateOffers} from '../mocks/offers.js';
 
 const TRIP_EVENT_COUNT = 20;
 
@@ -38,12 +38,12 @@ render(tripMenuElement, createEventFilterTemplate(), `beforeend`);
 const tripEventsElement = siteMainElement.querySelector(`.trip-events`);
 render(tripEventsElement, createEventSorterTemplate(), `beforeend`);
 
-render(tripEventsElement, createEventFormTemplate(tripEvents[0]), `beforeend`);
+render(tripEventsElement, createEventEditorTemplate(tripEvents[0]), `beforeend`);
 
 render(tripEventsElement, createEventDayTemplate(), `beforeend`);
 
 const tripDayElement = tripEventsElement.querySelector(`.trip-events__list`);
 
 for (let i = 1; i < TRIP_EVENT_COUNT; i++) {
-  render(tripDayElement, createEventItemTemplate(tripEvents[i]), `beforeend`);
+  render(tripDayElement, createEventPointTemplate(tripEvents[i]), `beforeend`);
 }
