+++ src/utils/trip.js
@@ -1,4 +1,4 @@
-import {ActivityType, PretextType} from '../src/const.js';
+import {ActivityType, PretextType} from '../const.js';
 
 export const pickEventPretext = (eventType) => {
   return Object.values(ActivityType).includes(eventType)
